refactor(NavState): move out of component
