+++ app/services/navigation/navState.ts
@@ -1,0 +1,7 @@
+export enum NavState {
+  DoneDisabled,
+  Done,
+  Current,
+  Open,
+  OpenDisabled,
+}
